renamed icon Link to NewWindow
