+++ src/Icon/NewWindow.js
